feat: setup different behaviours with strategy pattern
