+++ src/services/OpenKlantRegistrationService/Listener/OpenKlantRegistrationHandler.ts
@@ -1,9 +1,8 @@
-import { Response } from '@gemeentenijmegen/apigateway-http';
+import { ApiGatewayV2Response, Response } from '@gemeentenijmegen/apigateway-http';
 import { ICatalogiApi } from './CatalogiApi';
 import { Notification } from './model/Notification';
-import { OpenKlantDigitaalAdresWithUuid } from './model/Partij';
 import { IOpenKlantApi } from './OpenKlantApi';
-import { OpenKlantMapper } from './OpenKlantMapper';
+import { RegistrationStrategyFactory } from './strategies/RegistrationStrategyFactory';
 import { IZakenApi } from './ZakenApi';
 
 export interface OpenKlantRegistrationServiceProps {
@@ -17,96 +16,32 @@ export interface OpenKlantRegistrationServiceProps {
 export class OpenKlantRegistrationHandler {
 
   private readonly configuration: OpenKlantRegistrationServiceProps;
+  private readonly strategyFactorty: RegistrationStrategyFactory;
   constructor(configuration: OpenKlantRegistrationServiceProps) {
     this.configuration = configuration;
-
+    this.strategyFactorty = new RegistrationStrategyFactory(this.configuration);
     if (process.env.DEBUG === 'true') {
       console.log('Configured using:', this.configuration);
     }
   }
 
-
-  async handleNotification(notification: Notification) {
-
+  async handleNotification(notification: Notification) : Promise<ApiGatewayV2Response> {
     if (process.env.DEBUG === 'true') {
       console.log('Recevied notification', notification);
     }
 
+    // Get the strategy that is in use currently
+    const strategy = this.strategyFactorty.buildStrategy();
+
     // Validate notification
-    const errors = this.validateNotification(notification);
+    const errors = strategy.validateNotification(notification);
     if (errors) {
-      // TODO return 200 when notification is of different type and not handled
-      return Response.json({ errors }, 200); // TODO change back in 400 later maybe?
+      return Response.json({ errors }, 206); // TODO change back in 400 later maybe?
     }
 
-    // Get the involved rol details and check if the role is the 'aanvrager'
-    const rolUrl = notification.resourceUrl;
-    const rol = await this.configuration.zakenApi.getRol(rolUrl);
-    const rolType = await this.configuration.catalogiApi.getRolType(rol.roltype);
-
-    // Check if role is of the target role type, otherwise return 200 but do not handle the notification
-    const isTargetRolType = this.configuration.roltypesToRegister.includes(rolType.omschrijvingGeneriek.toLocaleLowerCase());
-    if (!isTargetRolType) {
-      console.debug('Role is not of the type to forward to open klant. Ignoring this notification.');
-      return Response.ok();
-    }
-    console.debug('Found a rol of the target type to forward to open klant.');
-
-    // Create a partij
-    const partijInput = OpenKlantMapper.partijFromRol(rol);
-    const partij = await this.configuration.openKlantApi.registerPartij(partijInput);
-    console.debug('Partij created', partij);
-
-    // Create a partij identificatie
-    const partijIdentificatieInput = OpenKlantMapper.partijIdentificatieFromRol(rol, partij.uuid);
-    const partijIdentificatie = await this.configuration.openKlantApi.addPartijIdentificatie(partijIdentificatieInput);
-    console.debug('Partij identificatie created', partijIdentificatie);
-
-    // Attach digitale adressen to partij
-    const digitaleAdressenInput = OpenKlantMapper.digitaalAdressenFromRol(rol, partij.uuid);
-    const promises: Promise<OpenKlantDigitaalAdresWithUuid>[] = [];
-    digitaleAdressenInput.forEach(digitaalAdres => {
-      promises.push(this.configuration.openKlantApi.addDigitaalAdres(digitaalAdres));
-    });
-    const digitaleAdressen = await Promise.all(promises);
-    digitaleAdressen.forEach(adres => console.log('Digitaal adres created', adres));
-
-    // Store the first digitaal adres as the prefered
-    // TODO figure out what the primary should be of the returned adressen?
-    const telefoon = digitaleAdressen.find(adres => adres.soortDigitaalAdres == OpenKlantMapper.TELEFOONNUMMER);
-    const email = digitaleAdressen.find(adres => adres.soortDigitaalAdres == OpenKlantMapper.EMAIL);
-    const voorkeur = telefoon?.uuid ?? email?.uuid;
-    if (!voorkeur) {
-      throw Error('No telefoon or email adres registered.');
-    }
-    await this.configuration.openKlantApi.updatePartij({
-      uuid: partij.uuid,
-      soortPartij: partij.soortPartij,
-      voorkeursDigitaalAdres: {
-        uuid: voorkeur,
-      },
-    });
-
-    return Response.ok();
-
+    // Handle notification as defined by the strategy
+    return strategy.register(notification);
   }
 
-
-  private validateNotification(notification: Notification) {
-    const errors: string[] = [];
-
-    if (notification.actie !== 'create' || notification.resource !== 'rol') {
-      errors.push(`Only rol creation notifications are handled by this endpoint (recevied: ${notification.actie}, ${notification.resource}).`);
-    }
-
-    if (!notification.hoofdObject.includes(this.configuration.zakenApiUrl)) {
-      errors.push('Notification points to a different ZRC than is configured for this endpoint.');
-    }
-
-    if (errors.length > 0) {
-      console.error('Notification validation failed', errors);
-    }
-    return errors.length == 0 ? undefined : errors;
-  }
 
 }
